requirements.txt
google-adk
mcp
playwright
google-genai
litellm
opik
python-dotenv
fastapi
uvicorn
